backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ML and NLP
torch==2.1.0
torch-geometric==2.4.0
sentence-transformers==2.2.2
transformers==4.35.0
numpy==1.24.3
scipy==1.11.4

# NLP preprocessing
spacy==3.7.2
nltk==3.8.1

# Authentication
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
typer==0.9.0
scikit-learn==1.3.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.1

